Move sha1hex to digest package
