+++ modules/core/shared/src/main/scala/ciris/Secret.scala
@@ -1,6 +1,6 @@
 package ciris
 
-import scala.annotation.tailrec
+import ciris.internal.digest.sha1Hex
 
 /**
   * [[Secret]] is used to denote that a configuration value is secret, and
@@ -50,7 +50,7 @@ final class Secret[A] private (val value: A) {
     * @return the SHA1 hex hash of `value.toString` in UTF-8
     */
   def valueHash: String =
-    Secret.sha1Hex(value.toString)
+    sha1Hex(value.toString)
 
   /**
     * Short version of [[valueHash]]. Only the first 7 characters are
@@ -94,33 +94,4 @@ object Secret {
 
   def unapply[A](secret: Secret[A]): Option[A] =
     Some(secret.value)
-
-  private[this] val hexChars: Array[Char] =
-    Array('0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f')
-
-  private def sha1Hex(s: String): String = {
-    def sha1(bytes: Array[Byte]): Array[Byte] = {
-      val digest = new internal.digest.SHA1Digest()
-      val hash = new Array[Byte](digest.digestSize)
-      digest.update(bytes, 0, bytes.length)
-      digest.doFinal(hash, 0)
-      hash
-    }
-
-    def hex(in: Array[Byte]): Array[Char] = {
-      val length = in.length
-
-      @tailrec def encode(out: Array[Char], i: Int, j: Int): Array[Char] = {
-        if (i < length) {
-          out(j) = hexChars((0xf0 & in(i)) >>> 4)
-          out(j + 1) = hexChars(0x0f & in(i))
-          encode(out, i + 1, j + 2)
-        } else out
-      }
-
-      encode(new Array(length << 1), 0, 0)
-    }
-
-    new String(hex(sha1(s.getBytes("UTF-8"))))
-  }
 }
